+++ fio.h
@@ -1006,7 +1006,8 @@ static inline void dprint_io_u(struct io_u *io_u, const char *p)
 	struct fio_file *f = io_u->file;
 
 	dprint(FD_IO, "%s: io_u %p: off=%llu/len=%lu/ddir=%d", p, io_u,
-					io_u->offset, io_u->buflen, io_u->ddir);
+					(unsigned long long) io_u->offset,
+					io_u->buflen, io_u->ddir);
 	if (fio_debug & (1 << FD_IO)) {
 		if (f)
 			log_info("/%s", f->file_name);
